create new rayTracer class and file
