+++ source/pfcClass.py
@@ -16,707 +16,13 @@ import multiprocessing
 import psutil
 import open3d as o3d
 from pickle import dump, load
-import drjit as dr
-import mitsuba as mi
+
+from rayTracerClass import shadowKernels
 
 tools = toolsClass.tools()
 
 import logging
 log = logging.getLogger(__name__)
-
-#class for ray triangle intersections to find shadows
-class shadowKernels:
-    def __init__(self):     
-        return 
-
-    def buildOpen3Dscene(self, targets):
-        """
-        builds a scene for open3D
-        """
-
-        Nt = len(targets)
-        #cast variables to 32bit for C
-        vertices = np.array(targets.reshape(Nt*3,3), dtype=np.float32)
-        triangles = np.array(np.arange(Nt*3).reshape(Nt,3), dtype=np.uint32)
-
-        #build intersection mesh and tensors for open3d
-        mesh = o3d.t.geometry.TriangleMesh()
-        scene = o3d.t.geometry.RaycastingScene()
-        mesh_id = scene.add_triangles(vertices, triangles)
-        self.scene = scene
-        return
-
-    def intersectTestOpen3D(self,q1,q2,targets,targetNorms, batchSize=1000):
-        """
-        checks if any of the lines (field line traces) generated by MAFOT
-        struct program intersect any of the target mesh faces.
-
-        Uses Open3D to accelerate the calculation:
-        Zhou, Qian-Yi, Jaesik Park, and Vladlen Koltun. "Open3D: A modern
-        library for 3D data processing." arXiv preprint arXiv:1801.09847 (2018).
-        """
-        t0 = time.time()
-        N = len(q2)
-        Nt = len(targets)
-        print('{:d} Source Faces and {:d} Target Faces in PFC object'.format(N,Nt))
-
-        mask = np.ones((N))
-
-        #construct rays
-        r = q2-q1
-        rMag = np.linalg.norm(r, axis=1)
-        rNorm = r / rMag.reshape((-1,1))
-
-        #calculate size of potential arrays in RAM
-        #availableRAM = psutil.virtual_memory().available #bytes
-
-        #calculate intersections
-        rays = o3d.core.Tensor([np.hstack([np.float32(q1),np.float32(rNorm)])],dtype=o3d.core.Dtype.Float32)
-        hits = self.scene.cast_rays(rays)
-        #convert open3d CPU tensors back to numpy
-        hitMap = hits['primitive_ids'][0].numpy()
-        distMap = hits['t_hit'][0].numpy()
-
-        #escapes occur where we have 32 bits all set: 0xFFFFFFFF = 4294967295 base10
-        escapes = np.where(hitMap == 4294967295)[0]
-        mask[escapes] = 0.0
-
-        #when distances to target exceed the trace step, we exclude any hits
-        tooLong = np.where(distMap > rMag)[0]
-        mask[tooLong] = 0.0
-
-        print('Found {:f} shadowed faces'.format(np.sum(mask)))
-        log.info('Found {:f} shadowed faces'.format(np.sum(mask)))
-        print('Time elapsed: {:f}'.format(time.time() - t0))
-        log.info('Time elapsed: {:f}'.format(time.time() - t0))
-
-        return mask
-
-
-    def getBatchSizeCUDA(self, N_rays):
-        """
-        Calculates a batch size for HEAT GPU Ray Casting.
-        
-        Logic:
-        We are launching N rays. The Scene (Targets) is static and already in VRAM.
-        We only need to ensure the RAYS don't exceed remaining memory.
-        
-        Memory per Ray approx:
-        - Origin (3 float) + Dir (3 float) + Mask (1 float) + Output (1 float)
-        - Mitsuba internal state (BVH stack, etc.) ~ 64-128 bytes
-        - DrJit Overhead
-        
-        Conservative estimate: 512 bytes per ray.
-        """
-        # 1. Get Free Memory
-        try:
-            free_mb = tools.get_free_gpu_memory_mb(0)
-        except:
-            # Fallback if tools isn't available
-            free_mb = 4096 
-            
-        print(f"[HEAT] Free GPU memory: {free_mb} MB")
-
-        # 2. Reserve Memory
-        # Reserve 1GB for the Scene Geometry + System Overhead
-        overhead_mb = 1024 
-        free_bytes = max(0, (free_mb - overhead_mb) * 1024**2)
-
-        # 3. Calculate Batch Size
-        # Conservative estimate: 512 bytes per ray is usually safe for simple intersection
-        bytes_per_ray = 512 
-        
-        if free_bytes <= 0:
-            batch_size = 10_000 # Fallback safety
-        else:
-            batch_size_est = int(free_bytes / bytes_per_ray)
-            # Cap at N_rays (don't need more than we have)
-            # Cap at 50 Million (prevent TDR timeouts on Windows/long kernels)
-            batch_size = max(1, min(N_rays, batch_size_est, 50_000_000))
-
-        print(f"[HEAT] Auto GPU batch size: {batch_size}")
-        return batch_size
-
-
-    def opticalShadows(self, use, centers, intersects, powerDir, MHD, CAD,
-                            shadowed_mask, controlfilePath, controlfileStruct, gridfileStruct, structOutfile,
-                            targetPoints, targetNorms, targetCtrs,
-                            t, ep, rtm, shadowMaskClouds=False):
-        """
-        Find shadowed faces for a given PFC object using MAFOT structure.
-        Traces field lines from PFC surface, looking for intersections with
-        triangles from intersectList meshes
-
-        rtm is rayTriMode, which defines the render engine used for ray tracing
-          Options are {mitsuba_gpu, mitsuba_cpu, open3d, heat}.  
-          'mitsuba_gpu' requires cuda installation.  'heat' is legacy and slow
-        """
-        print("\nFinding intersections for {:d} faces".format(len(centers[use])))
-        log.info("\nFinding intersections for {:d} faces".format(len(centers[use])))
-        print('Number of target parts: {:f}'.format(len(intersects)))
-        log.info('Number of target parts: {:f}'.format(len(intersects)))
-
-        #build the target mesh
-        r,z,phi = tools.xyz2cyl(targetCtrs[:,0],targetCtrs[:,1],targetCtrs[:,2])
-        targetBNorms = MHD.Bfield_pointcloud(ep, r, z, phi, powerDir=None, normal=True)
-        bdotnTgt = np.multiply(targetNorms, targetBNorms).sum(1)
-        powerDirTgt = tools.calculatePowerDir(bdotnTgt, ep.g['Bt0'])
-        fwdUseTgt = np.where(powerDirTgt > 0)[0]
-        revUseTgt = np.where(powerDirTgt < 0)[0]
-
-        #for debugging, save a shadowmask at each step up fieldline
-        if shadowMaskClouds == True:
-            IO = ioClass.IO_HEAT(self.chmod, self.UID, self.GID)
-            IO.writePointCloudCSV(centers,shadowed_mask,controlfilePath,'shadowMask',None,'shadowMask')
-
-        if rtm == 'mitsuba_gpu':
-            mi.set_variant('cuda_ad_rgb')
-        elif rtm == 'mitsuba_cpu':
-            mi.set_variant('llvm_ad_rgb')
-
-        #===INTERSECTION TEST 1 (tricky frontface culling / first step up field line)
-        dphi = 1.0
-        MHD.ittStruct = 1.0
-        numSteps = MHD.nTrace #actual trace is (numSteps + 1)*dphi/dpinit degrees
-        #If numSteps = 0, dont do intersection checking
-        if numSteps > 0:
-            print("\n----Intersection Step 1----")
-            log.info("\n----Intersection Step 1----")
-            CTLfile = controlfilePath + controlfileStruct
-            #q1 = np.zeros((len(self.centers[use]),3))
-            #q2 = np.zeros((len(self.centers[use]),3))
-
-            #run forward mesh elements
-            print("-Forward Trace-")
-            log.info("-Forward Trace-")
-            mapDirectionStruct = 1.0
-            startIdx = 1 #Match MAFOT sign convention for toroidal direction (CCW=+)
-            fwdUse = np.where(powerDir[use]==-1)[0]
-            if len(fwdUse) != 0:
-                MHD.writeControlFile(CTLfile, t, mapDirectionStruct, mode='struct')
-                #Perform first integration step
-                #MHD.writeMAFOTpointfile(self.centers[fwdUse],self.gridfileStruct)
-                MHD.writeMAFOTpointfile(centers[use[fwdUse]],gridfileStruct)
-                MHD.getMultipleFieldPaths(dphi, gridfileStruct, controlfilePath, controlfileStruct)
-                structData = tools.readStructOutput(structOutfile)
-                shutil.copyfile(structOutfile, structOutfile + '_fwdUse_step1')
-                os.remove(structOutfile)
-                q1 = structData[0::2,:] #even indexes are first trace point
-                q2 = structData[1::2,:] #odd indexes are second trace point
-                #build render engine scene and calculate intersections
-                if rtm == 'mitsuba_cpu':
-                    scene = self.buildMitsubaScene(targetPoints[fwdUseTgt], mitsubaMode='cpu')
-                    intersect_mask = self.intersectTestMitsuba(q1,q2,scene,mitsubaMode='cpu')
-                elif rtm == 'mitsuba_gpu':
-                    scene = self.buildMitsubaScene(targetPoints[fwdUseTgt], mitsubaMode='cuda')
-                    intersect_mask = self.intersectTestMitsuba(q1,q2,scene,mitsubaMode='cuda')
-                else: #use open3D
-                    scene = self.buildOpen3Dscene(targetPoints[fwdUseTgt])
-                    intersect_mask = self.intersectTestOpen3D(q1,q2,targetPoints[fwdUseTgt],targetNorms[fwdUseTgt], scene)
-                self.shadowed_mask[use[fwdUse]] = intersect_mask
-
-            #run reverse mesh elements
-            print("-Reverse Trace-")
-            log.info("-Reverse Trace-")
-            mapDirectionStruct = -1.0
-            startIdx = 0 #Match MAFOT sign convention for toroidal direction
-            revUse = np.where(powerDir[use]==1)[0]
-            if len(revUse) != 0:
-                MHD.writeControlFile(CTLfile, t, mapDirectionStruct, mode='struct')
-                #Perform first integration step
-                #MHD.writeMAFOTpointfile(self.centers[revUse],self.gridfileStruct)
-                MHD.writeMAFOTpointfile(centers[use[revUse]],gridfileStruct)
-                MHD.getMultipleFieldPaths(dphi, gridfileStruct, controlfilePath, controlfileStruct)
-                structData = tools.readStructOutput(structOutfile)
-                shutil.copyfile(structOutfile, structOutfile + '_revUse_step1')
-                os.remove(structOutfile)
-                q1 = structData[1::2,:] #odd indexes are first trace point
-                q2 = structData[0::2,:] #even indexes are second trace point
-                #build render engine scene and calculate intersections
-                if rtm == 'mitsuba_cpu':
-                    scene = self.buildMitsubaScene(targetPoints[revUseTgt], mitsubaMode='cpu')
-                    intersect_mask = self.intersectTestMitsuba(q1,q2,scene,mitsubaMode='cpu')
-                elif rtm == 'mitsuba_gpu':
-                    scene = self.buildMitsubaScene(targetPoints[revUseTgt], mitsubaMode='cuda')
-                    intersect_mask = self.intersectTestMitsuba(q1,q2,scene,mitsubaMode='cuda')
-                else: #use open3D
-                    scene = self.buildOpen3Dscene(targetPoints[revUseTgt])
-                    intersect_mask = self.intersectTestOpen3D(q1,q2,targetPoints[revUseTgt],targetNorms[revUseTgt], scene)
-                shadowed_mask[use[revUse]] = intersect_mask
-
-            #this is for printing information about a specific mesh element
-            #you can get the element # from paraview Point ID
-            #by default turned off
-            paraviewIndex = None
-            #paraviewIndex = 1
-            if paraviewIndex is not None:
-                ptIdx = np.where(use==paraviewIndex)[0]
-                print("Finding intersection face for point at:")
-                print(centers[paraviewIndex])
-            else:
-                ptIdx = None
-
-            #for debugging, save a shadowmask at each step up fieldline
-            if shadowMaskClouds == True:
-                IO.writePointCloudCSV(centers,shadowed_mask,controlfilePath,'shadowMask',None,'shadowMask')
-
-        #===INTERSECTION TEST 2 (multiple steps up field line)
-        #build scene once outside of loop
-        if rtm == 'mitsuba_cpu':
-            scene = self.buildMitsubaScene(targetPoints, mitsubaMode='cpu')
-        elif rtm == 'mitsuba_gpu':
-            scene = self.buildMitsubaScene(targetPoints, mitsubaMode='cuda')
-        else: #use open3D
-            scene = self.buildOpen3Dscene(targetPoints)
-
-        #Starts at second step up field line
-        if numSteps > 1:
-            print("\n----Intersection Step 2----")
-            log.info("\n----Intersection Step 2----")
-            use = np.where(shadowed_mask == 0)[0]
-            intersect_mask2 = np.zeros((len(use)))
-            q2 = np.zeros((len(centers[use]),3))
-
-            #calculate q2 for initialization of the big loop below
-            #run forward mesh elements
-            print("-Forward Trace-")
-            log.info("-Forward Trace-")
-            mapDirectionStruct = 1.0
-            startIdx = 1 #Match MAFOT sign convention for toroidal direction
-            fwdUse = np.where(powerDir[use]==-1)[0]
-            if len(fwdUse) != 0:
-                MHD.writeControlFile(CTLfile, t, mapDirectionStruct, mode='struct')
-                #Perform first integration step
-                MHD.writeMAFOTpointfile(centers[use[fwdUse]],gridfileStruct)
-                MHD.getMultipleFieldPaths(dphi, gridfileStruct, controlfilePath, controlfileStruct)
-                structData = tools.readStructOutput(structOutfile)
-                os.remove(structOutfile) #clean up
-                q2[fwdUse] = structData[1::2,:] #odd indexes are second trace point
-
-            #run reverse mesh elements
-            print("-Reverse Trace-")
-            log.info("-Reverse Trace-")
-            mapDirectionStruct = -1.0
-            startIdx = 0 #Match MAFOT sign convention for toroidal direction
-            revUse = np.where(powerDir[use]==1)[0]
-            if len(revUse) != 0:
-                MHD.writeControlFile(CTLfile, t, mapDirectionStruct, mode='struct')
-                #Perform first integration step
-                MHD.writeMAFOTpointfile(centers[use[revUse]],gridfileStruct)
-                MHD.getMultipleFieldPaths(dphi, gridfileStruct, controlfilePath, controlfileStruct)
-                structData = tools.readStructOutput(structOutfile)
-                os.remove(structOutfile) #clean up
-                q2[revUse] = structData[0::2,:] #even indexes are second trace point
-
-            if paraviewIndex is not None:
-                ptIdx = np.where(use==paraviewIndex)[0]
-                print("Finding intersection face for point at:")
-                print(centers[paraviewIndex])
-            else:
-                ptIdx = None
-
-            #Perform subsequent integration steps.  Use the point we left off at in
-            #last loop iteration as the point we launch from in next loop iteration
-            #This amounts to 'walking' up the field line looking for intersections,
-            #which is important when field line curvature makes intersections happen
-            #farther than 1-2 degrees from PFC surface.
-            #
-            #if you need to reference stuff from the original arrays (ie self.centers)
-            #you need to do nested uses (ie: self.centers[use][use2]).
-            use2 = np.where(intersect_mask2 == 0)[0]
-            for i in range(numSteps):
-                print("\n----Intersect Trace Step {:d}----".format(i+3))
-                log.info("\n----Intersect Trace Step {:d}----".format(i+3))
-
-                useOld = use2
-                use2 = np.where(intersect_mask2 == 0)[0]
-
-                #if all faces are shadowed, break
-                if len(use2) == 0:
-                    print("All faces shadowed on this PFC. Moving onto next PFC...")
-                    log.info("All faces shadowed on this PFC. Moving onto next PFC...")
-                    break
-
-                #map current steps's index back to intersect_mask2 index
-                indexes = np.intersect1d(use2,useOld, return_indices=True)[-1]
-                if paraviewIndex is not None:
-                    ptIdx = np.where(use2==useOld[ptIdx])[0] #for tracing intersection locations
-                else:
-                    ptIdx = None
-
-                StartPoints = q2[indexes].copy() #odd indexes are second trace point
-                q1 = np.zeros((len(StartPoints),3))
-                q2 = np.zeros((len(StartPoints),3))
-                #run forward mesh elements
-                print("-Forward Trace-")
-                log.info("-Forward Trace-")
-                mapDirectionStruct = 1.0
-                startIdx = 1 #Match MAFOT sign convention for toroidal direction
-                fwdUse = np.where(powerDir[use[use2]]==-1)[0]
-                if len(fwdUse) == 0:
-                    print("No more traces in forward direction")
-                    log.info("No more traces in forward direction")
-                else:
-                    MHD.writeControlFile(CTLfile, t, mapDirectionStruct, mode='struct')
-                    #Perform first integration step
-                    MHD.writeMAFOTpointfile(StartPoints[fwdUse],gridfileStruct)
-                    MHD.getMultipleFieldPaths(dphi, gridfileStruct, controlfilePath, controlfileStruct)
-                    structData = tools.readStructOutput(structOutfile)
-                    os.remove(structOutfile) #clean up
-                    q1[fwdUse] = structData[0::2,:] #odd indexes are second trace point
-                    q2[fwdUse] = structData[1::2,:] #odd indexes are second trace point
-                    insideBndy,insideBndyIdx = self.checkInsideBoundingBox(q2[fwdUse], MHD, CAD)
-                    checkMask2 = ~insideBndy     # field lines outside the Bounding box are shadowed
-                    #calculate intersections
-                    if rtm == 'mitsuba_cpu':
-                        checkMask2[insideBndyIdx] = self.intersectTestMitsuba(q1[fwdUse][insideBndyIdx],q2[fwdUse][insideBndyIdx],scene,mitsubaMode='cpu')
-                    elif rtm == 'mitsuba_gpu':
-                        checkMask2[insideBndyIdx]= self.intersectTestMitsuba(q1[fwdUse][insideBndyIdx],q2[fwdUse][insideBndyIdx],scene,mitsubaMode='cuda')
-                    else: #use open3D
-                        checkMask2[insideBndyIdx] = self.intersectTestOpen3D(q1[fwdUse][insideBndyIdx],q2[fwdUse][insideBndyIdx],targetPoints,targetNorms, scene)
- 
-                    #print('Verify the Box Mask Fwd: old =', np.sum(checkMask1), '  new =', np.sum(checkMask2), '  Okay?', np.sum(checkMask1) <= np.sum(checkMask2))
-                    intersect_mask2[use2[fwdUse]] = checkMask2
-
-                #run reverse mesh elements
-                print("-Reverse Trace-")
-                log.info("-Reverse Trace-")
-                mapDirectionStruct = -1.0
-                startIdx = 0 #Match MAFOT sign convention for toroidal direction
-                revUse = np.where(powerDir[use[use2]]==1)[0]
-                if len(revUse) == 0:
-                    print("No more traces in reverse direction")
-                    log.info("No more traces in reverse direction")
-                else:
-                    MHD.writeControlFile(CTLfile, t, mapDirectionStruct, mode='struct')
-                    #Perform first integration step
-                    MHD.writeMAFOTpointfile(StartPoints[revUse],gridfileStruct)
-                    MHD.getMultipleFieldPaths(dphi, gridfileStruct, controlfilePath, controlfileStruct)
-                    structData = tools.readStructOutput(structOutfile)
-                    os.remove(structOutfile) #clean up
-                    q1[revUse] = structData[1::2,:] #odd indexes are second trace point
-                    q2[revUse] = structData[0::2,:] #odd indexes are second trace point
-                    insideBndy,insideBndyIdx = self.checkInsideBoundingBox(q2[revUse], MHD, CAD)
-                    checkMask2 = ~insideBndy     # field lines outside the Bounding box are shadowed
-                    #calculate intersections
-                    if rtm == 'mitsuba_cpu':
-                        checkMask2[insideBndyIdx] = self.intersectTestMitsuba(q1[revUse][insideBndyIdx],q2[revUse][insideBndyIdx],scene,mitsubaMode='cpu')
-                    elif rtm == 'mitsuba_gpu':
-                        checkMask2[insideBndyIdx]= self.intersectTestMitsuba(q1[revUse][insideBndyIdx],q2[revUse][insideBndyIdx],scene,mitsubaMode='cuda')
-                    else: #use open3D
-                        checkMask2[insideBndyIdx] = self.intersectTestOpen3D(q1[revUse][insideBndyIdx],q2[revUse][insideBndyIdx],targetPoints,targetNorms, scene)
-
-                    #print('Verify the Box Mask Rev: old =', np.sum(checkMask1), '  new =', np.sum(checkMask2), '  Okay?', np.sum(checkMask1) <= np.sum(checkMask2))
-                    intersect_mask2[use2[revUse]] = checkMask2
-
-                #for debugging, save a shadowmask at each step up fieldline
-                if shadowMaskClouds == True:
-                    shadowed_mask[use] = intersect_mask2
-                    IO.writePointCloudCSV(centers,shadowed_mask,controlfilePath,'shadowMask','test{:d}'.format(i+1),'shadowMask')
-                print("Step {:d} complete".format(i))
-                log.info("Step {:d} complete".format(i))
-            #Now revise shadowed_mask taking intersections into account
-            shadowed_mask[use] = intersect_mask2
-        print("Completed Intersection Check")
-        return shadowed_mask
-
-
-
-    def opticalShadowsBatch(self, use, centers, intersects, powerDir, MHD, CAD,
-                            shadowed_mask, controlfilePath, controlfileStruct, gridfileStruct, structOutfile,
-                            targetPoints, targetNorms, targetCtrs,
-                            t, ep, rtm, shadowMaskClouds=False):
-        """
-        Does all rayTri intersection checking in a single pass rather than step by step
-        This is not always faster, as going step by step enables us to prune traces that
-        intersect as we go.
-        
-        Find shadowed faces for a given PFC object using MAFOT structure.
-        Traces field lines from PFC surface, looking for intersections with
-        triangles from intersectList meshes
-        """
-        print("\nFinding intersections for {:d} faces".format(len(self.centers[use])))
-        log.info("\nFinding intersections for {:d} faces".format(len(self.centers[use])))
-        print('Number of target parts: {:f}'.format(len(self.intersects)))
-        log.info('Number of target parts: {:f}'.format(len(self.intersects)))
-
-        #build the target mesh
-        r,z,phi = tools.xyz2cyl(targetCtrs[:,0],targetCtrs[:,1],targetCtrs[:,2])
-        targetBNorms = MHD.Bfield_pointcloud(self.ep, r, z, phi, powerDir=None, normal=True)
-        bdotnTgt = np.multiply(targetNorms, targetBNorms).sum(1)
-        powerDirTgt = tools.calculatePowerDir(bdotnTgt, self.ep.g['Bt0'])
-        fwdUseTgt = np.where(powerDirTgt > 0)[0]
-        revUseTgt = np.where(powerDirTgt < 0)[0]
-
-        #trace field lines
-        dphi = 1.0
-        MHD.ittStruct = MHD.nTrace #actual trace is (nTrace + 1)*dphi/dpinit degrees
-        if MHD.nTrace > 0:
-            print("\n----Tracing Field Lines----")
-            log.info("\n----Tracing Field Lines----")
-            CTLfile = self.controlfilePath + self.controlfileStruct
-
-            q1_blocks, q2_blocks, q1_0, q2_0, face_idx_blocks = [], [], [], [], []
-            #run forward mesh elements
-            print("-Forward Trace-")
-            log.info("-Forward Trace-")
-            mapDirectionStruct = 1.0
-            fwdUse = np.where(self.powerDir[use]==-1)[0]
-            if len(fwdUse) != 0:
-                #run full MAFOT trace
-                MHD.writeControlFile(CTLfile, self.t, mapDirectionStruct, mode='struct')
-                MHD.writeMAFOTpointfile(self.centers[use[fwdUse]],self.gridfileStruct)
-                MHD.getMultipleFieldPaths(dphi, self.gridfileStruct, self.controlfilePath, self.controlfileStruct)
-                structDataFwd = tools.readStructOutput(self.structOutfile)
-                os.remove(self.structOutfile) #clean up
-                structData = structDataFwd.reshape(len(fwdUse), MHD.nTrace+1, 3)
-                #tricky front face culling for 1st field line step
-                q1Fwd0 = structData[:, 0, :].reshape(-1, 3)
-                q2Fwd0 = structData[:, 1, :].reshape(-1, 3)
-
-                #build render engine scene and calculate intersections
-                if rtm == 'mitsuba_cpu':
-                    scene = self.buildMitsubaScene(targetPoints[fwdUseTgt], mitsubaMode='cpu')
-                    self.shadowed_mask[use[fwdUse]] = self.intersectTestMitsuba(q1Fwd0,q2Fwd0,scene,mitsubaMode='cpu')
-                elif rtm == 'mitsuba_gpu':
-                    scene = self.buildMitsubaScene(targetPoints[fwdUseTgt], mitsubaMode='cuda')
-                    self.shadowed_mask[use[fwdUse]] = self.intersectTestMitsuba(q1Fwd0,q2Fwd0,scene,mitsubaMode='cuda')
-                else: #use open3D
-                    scene = self.buildOpen3Dscene(targetPoints[fwdUseTgt])
-                    self.shadowed_mask[use[fwdUse]] = self.intersectTestOpen3D(q1Fwd0,q2Fwd0,targetPoints[fwdUseTgt],targetNorms[fwdUseTgt], scene)
-
-                #all remaining trace steps in single array
-                q1Fwd = structData[:, 1:MHD.nTrace, :]
-                q2Fwd = structData[:, 2:MHD.nTrace+1, :]
-                q1_blocks.append(q1Fwd.reshape(-1, 3)) 
-                q2_blocks.append(q2Fwd.reshape(-1, 3)) 
-                face_idx_blocks.append(use[fwdUse])
-
-            #run reverse mesh elements
-            print("-Reverse Trace-")
-            log.info("-Reverse Trace-")
-            mapDirectionStruct = -1.0
-            revUse = np.where(self.powerDir[use]==1)[0]
-            if len(revUse) != 0:
-                #run full MAFOT trace
-                MHD.writeControlFile(CTLfile, self.t, mapDirectionStruct, mode='struct')
-                MHD.writeMAFOTpointfile(self.centers[use[revUse]],self.gridfileStruct)
-                MHD.getMultipleFieldPaths(dphi, self.gridfileStruct, self.controlfilePath, self.controlfileStruct)
-                structDataRev = tools.readStructOutput(self.structOutfile)
-                os.remove(self.structOutfile) #clean up
-                structData = structDataRev.reshape(len(revUse), MHD.nTrace+1, 3)
-                structData = np.flip(structData, axis=1) #flip data because these are reverse traces
-                #tricky front face culling for 1st field line step
-                q1Rev0 = structData[:, 0, :].reshape(-1, 3)
-                q2Rev0 = structData[:, 1, :].reshape(-1, 3)
-
-                #build render engine scene and calculate intersections
-                if rtm == 'mitsuba_cpu':
-                    scene = self.buildMitsubaScene(targetPoints[revUseTgt], mitsubaMode='cpu')
-                    self.shadowed_mask[use[revUse]] = self.intersectTestMitsuba(q1Rev0,q2Rev0,scene,mitsubaMode='cpu')
-                elif rtm == 'mitsuba_gpu':
-                    scene = self.buildMitsubaScene(targetPoints[revUseTgt], mitsubaMode='cuda')
-                    self.shadowed_mask[use[revUse]] = self.intersectTestMitsuba(q1Rev0,q2Rev0,scene,mitsubaMode='cuda')
-                else: #use open3D
-                    scene = self.buildOpen3Dscene(targetPoints[revUseTgt])
-                    self.shadowed_mask[use[revUse]] = self.intersectTestOpen3D(q1Rev0,q2Rev0,targetPoints[revUseTgt],targetNorms[revUseTgt], scene)
-
-                q1Rev = structData[:, 1:MHD.nTrace, :]
-                q2Rev = structData[:, 2:MHD.nTrace+1, :] 
-                q1_blocks.append(q1Rev.reshape(-1, 3)) 
-                q2_blocks.append(q2Rev.reshape(-1, 3))
-                face_idx_blocks.append(use[revUse])
-
-            # concatenate all available traces
-            q1_all = np.concatenate(q1_blocks, axis=0)
-            q2_all = np.concatenate(q2_blocks, axis=0)
-            face_idx_all = np.concatenate(face_idx_blocks, axis=0)
-            seg2seed = np.repeat(face_idx_all, MHD.nTrace-1) #maps trace steps (segments) to PFC centroid indexes (seeds) (subtract one for first steps)
-
-            # batch intersection calculation for entire trace
-            #build scene once outside of loop
-            if rtm == 'mitsuba_cpu':
-                scene = self.buildMitsubaScene(targetPoints, mitsubaMode='cpu')
-                ray_mask = self.intersectTestMitsuba(q1_all,q2_all,scene,mitsubaMode='cpu')
-            elif rtm == 'mitsuba_gpu':
-                scene = self.buildMitsubaScene(targetPoints, mitsubaMode='cuda')
-                ray_mask = self.intersectTestMitsuba(q1_all,q2_all,scene,mitsubaMode='cuda')
-            else: #use open3D
-                scene = self.buildOpen3Dscene(targetPoints)
-                ray_mask = self.intersectTestOpen3D(q1_all, q2_all, targetPoints, targetNorms, scene)
-
-            # --- Reduce per-seed: a seed is shadowed if ANY of its segments hit ---
-            # Use indexed accumulation, then threshold
-            N_faces = self.shadowed_mask.shape[0]
-            acc = np.zeros(N_faces, dtype=np.int32)           
-            np.add.at(acc, seg2seed, ray_mask.astype(np.int8))
-            seed_hit = acc > 0  # boolean per global face index
-            self.shadowed_mask[face_idx_all] += seed_hit[face_idx_all].astype(self.shadowed_mask.dtype)
-            self.shadowed_mask[self.shadowed_mask > 0] = 1 #convert back to bool
-        print("Completed Intersection Check")
-        return
-
-
-
-
-    def buildMitsubaScene(self, targets, mitsubaMode='cuda', dedup_tol=1e-9):
-        """
-        Mitsuba scene build from triangle vertices using direct Mesh instantiation.
-
-        - Deduplicates vertices
-        - Uploads float32 vertices and uint32 faces
-        - Returns a scene with one mesh named 'targets'
-        
-        Parameters
-        ----------
-        targets : (Nt,3,3) np.ndarray
-            Triangle vertices in mm (PFC coordinates). 
-            NOTE: This function converts them to Meters internally.
-        dedup_tol : float
-            Quantization tolerance.
-        """
-        # 1. Set Variant Safely
-        current_variant = getattr(mi, 'variant', lambda: None)()
-        target_variant = 'cuda_ad_rgb' if mitsubaMode == 'cuda' else 'llvm_ad_rgb'
-        if current_variant != target_variant:
-            try:
-                mi.set_variant(target_variant)
-            except Exception as e:
-                print(f"[HEAT] Warning: Could not switch variant to {target_variant}. Current: {current_variant}")
-
-        # 2. Prepare Data (Numpy CPU Side)
-        targets = np.asarray(targets)
-        if targets.ndim != 3 or targets.shape[1:] != (3, 3):
-            raise ValueError("targets must have shape (Nt, 3, 3)")
-
-        Nt = targets.shape[0]
-        verts = targets.reshape(Nt * 3, 3)
-
-        # Deduplicate
-        q = np.round(verts / dedup_tol).astype(np.int64)
-        uq, inv = np.unique(q, axis=0, return_inverse=True)
-        
-        # Recover vertices
-        # We do this here because we can't easily apply 'to_world' 
-        # to a raw Mesh object in the dictionary without a wrapping Instance.
-        vertices_np = (uq.astype(np.float64) * dedup_tol).astype(np.float32)
-        faces_np = inv.reshape(Nt, 3).astype(np.uint32)
-
-        # 3. Create the Mitsuba Mesh Object
-        # We set has_vertex_normals=False to ensure we use geometric (face) normals
-        # (Equivalent to 'face_normals': True in the PLY loader)
-        mesh = mi.Mesh("targets",
-                       vertex_count=len(vertices_np),
-                       face_count=len(faces_np),
-                       has_vertex_normals=False,
-                       has_vertex_texcoords=False)
-
-        # 4. Upload Data to GPU/DrJit
-        params = mi.traverse(mesh)
-        
-        # Flatten and cast to appropriate Mitsuba types
-        params['vertex_positions'] = dr.ravel(mi.Float(vertices_np.flatten()))
-        params['faces'] = dr.ravel(mi.UInt(faces_np.flatten()))
-        params.update()
-
-        # 5. Build Scene
-        # We pass the actual python mesh object into the dictionary
-        scene_dict = {
-            'type': 'scene',
-            'targets': mesh
-        }
-        
-        scene = mi.load_dict(scene_dict)
-        self.scene = scene
-        return scene
-
-    def intersectTestMitsuba(self,
-                             q1,
-                             q2,
-                             scene,
-                             mitsubaMode='cuda'):
-        """
-        Calculates intersections using Mitsuba render engine.
-        Optimized for Shadow/Occlusion checks (ray_test).
-        
-        Parameters
-        ----------
-        q1, q2 : (N, 3) np.ndarray
-            Start and end points of the segment along the field line [m].
-        scene : mitsuba.Scene
-            Built by buildMitsubaScene.
-        mitsubaMode : {'cuda', 'llvm'}
-    
-        Returns
-        -------
-        mask : (N,) np.ndarray
-            1.0 where the segment is shadowed (hit within |q2 - q1|), 0.0 otherwise.
-        """
-        batchSize = self.getBatchSizeCUDA(len(q1))
-
-        # Ensure correct variant is active
-        # (This check is cheap if already active)
-        target_variant = 'cuda_ad_rgb' if mitsubaMode == 'cuda' else 'llvm_ad_rgb'
-        print(f"[HEAT] Using ray tracer variant: {target_variant}")
-        if getattr(mi, 'variant', lambda: None)() != target_variant:
-            try:
-                mi.set_variant(target_variant)
-            except Exception as e:
-                print(f"[HEAT] Warning: Variant switch issue in intersectTest: {e}")
-
-        Float = mi.Float
-    
-        q1 = np.asarray(q1, dtype=np.float64)
-        q2 = np.asarray(q2, dtype=np.float64)
-        N  = q1.shape[0]
-        # Result mask (1.0 = shadowed, 0.0 = escape)
-        mask = np.zeros((N,), dtype=np.float32)
-    
-        # Precompute segment direction and length
-        r     = q2 - q1
-        rMag  = np.linalg.norm(r, axis=1)          # (N,)
-        rNorm = r / rMag.reshape(-1, 1)            # (N, 3)
-    
-        eps = 1e-6  # Epsilon to avoid self-intersection at origin
-    
-        for start in range(0, N, batchSize):
-            end  = min(start + batchSize, N)
-            
-            # Get Batch Slices (Numpy)
-            q1_b    = q1[start:end]
-            rNorm_b = rNorm[start:end]
-            rMag_b  = rMag[start:end]
-    
-            # Convert to DrJit Arrays
-            # Explicit Float() cast is crucial to avoid Backend Mismatch with numpy arrays
-            o = mi.Point3f(Float(q1_b[:, 0]),
-                           Float(q1_b[:, 1]),
-                           Float(q1_b[:, 2]))
-            
-            d = mi.Vector3f(Float(rNorm_b[:, 0]),
-                            Float(rNorm_b[:, 1]),
-                            Float(rNorm_b[:, 2]))
-            
-            dist = Float(rMag_b)
-            ray = mi.Ray3f(o=o, d=d)
-            ray.maxt=dist
-            #ray.mint = Float(eps)
-    
-            # ray_test() returns a boolean mask if ANY hit occurs between mint and maxt.
-            # This is much faster than ray_intersect() for occlusion checks.
-            is_shadowed = scene.ray_test(ray)
-    
-            # Convert Bool mask to Float (1.0 or 0.0)
-            mask_b = dr.select(is_shadowed, Float(1.0), Float(0.0))
-            
-            # Evaluate and Copy back to CPU
-            # dr.eval enforces computation before the copy
-            dr.eval(mask_b)
-            mask[start:end] = np.array(mask_b, copy=False)
-    
-        return mask
-
-
 
 
 class PFC(shadowKernels):
