+++ source/radClass.py
@@ -14,6 +14,8 @@ import psutil
 #from scipy.spatial import ConvexHull
 
 import toolsClass
+from rayTracerClass import shadowKernels, triangle_centroids_and_areas
+
 log = logging.getLogger(__name__)
 tools = toolsClass.tools()
 
@@ -194,6 +196,10 @@ class RAD:
         prepares the radiation sources and targets (ROI PFCs) for the
         radiated power calc.  Gets the source centers, target centers, and
         intersection mesh vertices.
+
+        For mergedPFCs, ROI triangles are taken from each PFClist child in list
+        order (matching concatenated centers), then occluders; child parts are
+        not re-added from intersectMeshes.
         """
         #build out source centers
         x,y,z = tools.cyl2xyz(self.PC3D[:,0,:].flatten(),
@@ -209,23 +215,38 @@ class RAD:
         self.targetAreas = PFC.areas
         targetMeshes = []
 
+        # mergedPFCs uses name 'mergedPFCs'; ROI rows in CAD use child part names.
+        # Match merged center order: vstack([p.centers for p in PFClist]).
+        child_names = set()
+        if getattr(PFC, 'mergedPFCs', False):
+            child_names = {p.name for p in PFC.PFClist}
+
         #build out intersect array
         numTargetFaces = 0
         intersectPoints = []
         intersectNorms = []
         totalMeshCounter = 0
 
-        #add ROI mesh
-        for i,target in enumerate(CAD.ROImeshes):
-            if CAD.ROIList[i] == PFC.name:
-                targetMeshes.append(target)
-                totalMeshCounter+=target.CountFacets
-                numTargetFaces += target.CountFacets
-                #append target data
-                for face in target.Facets:
-                    intersectPoints.append(face.Points)
-                    intersectNorms.append(face.Normal)
-
+        #add ROI mesh (faces must align with PFC.centers facet order)
+        if getattr(PFC, 'mergedPFCs', False):
+            for child in PFC.PFClist:
+                for i, target in enumerate(CAD.ROImeshes):
+                    if CAD.ROIList[i] == child.name:
+                        targetMeshes.append(target)
+                        totalMeshCounter += target.CountFacets
+                        numTargetFaces += target.CountFacets
+                        for face in target.Facets:
+                            intersectPoints.append(face.Points)
+                            intersectNorms.append(face.Normal)
+        else:
+            for i, target in enumerate(CAD.ROImeshes):
+                if CAD.ROIList[i] == PFC.name:
+                    targetMeshes.append(target)
+                    totalMeshCounter += target.CountFacets
+                    numTargetFaces += target.CountFacets
+                    for face in target.Facets:
+                        intersectPoints.append(face.Points)
+                        intersectNorms.append(face.Normal)
 
         for i,intersect in enumerate(CAD.intersectMeshes):
             try:
@@ -239,10 +260,12 @@ class RAD:
                 #exclude self shadowing
                 if CAD.intersectList[i] == PFC.name:
                     pass
+                # merged ROI already appended above; skip duplicate from intersectMeshes
+                elif CAD.intersectList[i] in child_names:
+                    pass
                 else:
                     targetMeshes.append(intersect)
                     numTargetFaces += intersect.CountFacets
-                    #append target data
                     for face in intersect.Facets:
                         intersectPoints.append(face.Points)
                         intersectNorms.append(face.Normal)
@@ -259,17 +282,10 @@ class RAD:
         self.Ni = len(self.sources)
         self.Nj = len(self.targetCtrs)
 
-        #build mesh for mitsuba3
+        # Mitsuba photon tracing builds the scene in-memory from self.intersectPoints [m]
+        # (same convention as rayTracerClass.shadowKernels.buildMitsubaScene).
         if 'mitsuba' in self.rayTracer:
-            combinedMesh = CAD.createEmptyMesh()
-            for m in targetMeshes:
-                combinedMesh.addFacets(m.Facets)
-            oldMask = CAD.overWriteMask
-            CAD.overWriteMask = True
-            #mitsuba only reads PLYs!!!
-            CAD.writeMesh2file(combinedMesh, 'combinedMesh', path=CAD.STLpath, resolution='standard', fType='ply')
-            CAD.overWriteMask = oldMask
-            self.meshFile = CAD.STLpath + 'combinedMesh' + "___standard.ply"    
+            self.meshFile = None
 
         #build mesh for HEAT algs (old and slow)
         elif self.rayTracer=='heat':
@@ -310,6 +326,9 @@ class RAD:
         Maps power between sources and targets (ROI PFCs).  Uses Mitsuba3 to
         perform ray tracing.  Mitsuba3 can be optimized for CPU or GPU.
 
+        Scene geometry uses rayTracerClass.shadowKernels.buildMitsubaScene on
+        self.intersectPoints (meters), consistent with optical tracing.
+
         Uses DrJIT to store arrays and for math operations
 
         Code largely developed by A. Rosenthal (CFS)
@@ -341,9 +360,8 @@ class RAD:
         print("Available CPU memory: {:f} GB".format(memAvail / (1024**3)))  
 
 
-        print("Building radiation scene...")    
+        print("Building radiation scene (rayTracerClass mesh, meters)...")
 
-        # Initialize Mitsuba
         if mitsubaMode == 'cuda':
             print("Using GPU")
             mi.set_variant('cuda_ad_rgb')
@@ -354,76 +372,23 @@ class RAD:
         Float = mi.Float   # backend-appropriate float (cuda / llvm)
         UInt  = mi.UInt
 
-        scene = {'type': 'scene'}
-        scene['path_int'] = {
-            'type': 'path',
-            'max_depth': 1  # Maximum number of bounces
-        }
-
-        name = 'allMesh' #name of entire mesh we saved
-
-        scene[name] = {
-            'type': fType,
-            'filename': self.meshFile,
-            'to_world': mi.ScalarTransform4f().scale([0.001, 0.001, 0.001]),  # Convert from mm to m
-            'face_normals': True,  # This prevents smoothing of sharp-corners by discarding surface-normals. Useful for engineering CAD.
-            }
-
+        #initialize ray tracer object
+        rt = shadowKernels()
         t = time.time()
-        print('Loading Scene....')
-        scene = mi.load_dict(scene)
-        print('Time to load scene: ' + str(time.time() - t))
+        scene = rt.buildMitsubaScene(self.intersectPoints, mitsubaMode=mitsubaMode)
+        print('Time to build Mitsuba scene: ' + str(time.time() - t))
         print('\n')
 
         t = time.time()
-        # Convert the numpy arrays to Dr.Jit dynamic arrays
-        #tx,tx,tz,tp are source values for x,y,z,power
+        #points power flows from
         tx = Float(self.sources[:, 0])
         ty = Float(self.sources[:, 1])
         tz = Float(self.sources[:, 2])
         tp = Float(self.sourcePower)
 
-
-        # Access the mesh object
-        mesh = None
-        for shape in scene.shapes():
-            if shape.id() == name:  # Use the ID or name of your mesh
-                mesh = shape
-                break
-            
-        if not mesh:
-            raise ValueError("Mesh not found in the scene!")
-
-        # Get vertex positions and face indices
-        params = mi.traverse(scene)
-        vertices = np.array(dr.unravel(mi.Point3f, params[name+'.vertex_positions'])).reshape(3,-1).T
-        # Retrieve the face indices (triangle indices)
-        faces = np.array(params[name + '.faces'], dtype=np.int32).reshape(-1,3)
-        totFace = len(faces)
-
-        # Compute normals, centers, and areas
-        v0 = vertices[faces[:, 0]]
-        v1 = vertices[faces[:, 1]]
-        v2 = vertices[faces[:, 2]]
-
-        # Edge vectors
-        edge1 = v1 - v0
-        edge2 = v2 - v0
-
-        # Normals (unnormalized)
-        normals = np.cross(edge1, edge2)
-
-        # Areas (half the magnitude of the cross-product)
-        area = np.linalg.norm(normals, axis=1) * 0.5
-
-        # Normalize normals
-        normals = normals / np.linalg.norm(normals, axis=1, keepdims=True)
-
-        # Centers (mean of vertices)
-        center = (v0 + v1 + v2) / 3.0       
+        center, area = triangle_centroids_and_areas(self.intersectPoints)
+        totFace = center.shape[0]
         N_sources = len(tx)
-        N_targets = len(center) #all intersections, including those not in ROI
-        targetPower = np.zeros((N_targets))
 
         #dynamically allocate batch size based upon available memory
         if batch_size == "auto":
@@ -563,7 +528,7 @@ class RAD:
 
     def traceMitsuba(self, sx,sy,sz,tx,ty,tz,tp,sA,scene, mitsubaMode, totFace, size_i):
         """
-        function to do trace so that garbage collection happens
+        Legacy helper; scene should be from rayTracerClass.shadowKernels.buildMitsubaScene.
         """
         import drjit as dr
         import mitsuba as mi
@@ -603,6 +568,8 @@ class RAD:
         THIS IS A LEGACY FUNCTION.  To use, need to update to be similar to JIT equivalent function,
         calculatePowerTransferMitsubaJIT.  Left for reference.
 
+        fType and meshFile are unused; scene is built from self.intersectPoints (meters).
+
         Maps power between sources and targets (ROI PFCs).  Uses Mitsuba3 to
         perform ray tracing.  Mitsuba3 can be optimized for CPU or GPU.
 
@@ -630,109 +597,36 @@ class RAD:
         #self.hullPower = np.zeros((self.Ni))
         #self.powerFrac = np.zeros((self.Ni))
 
-        print("Building radiation scene...")    
+        print("Building radiation scene (rayTracerClass mesh, meters)...")
 
-        # Initialize Mitsuba
         if mitsubaMode == 'cuda':
             mi.set_variant('cuda_ad_rgb')
         else:
             mi.set_variant('llvm_ad_rgb')
 
-        scene = {'type': 'scene'}
-        scene['path_int'] = {
-            'type': 'path',
-            'max_depth': 1  # Maximum number of bounces
-        }
-
-        name = 'allMesh' #name of entire mesh we saved
-
-        scene[name] = {
-            'type': fType,
-            'filename': self.meshFile,
-            'to_world': mi.ScalarTransform4f().scale([0.001, 0.001, 0.001]),  # Convert from mm to m
-            'face_normals': True,  # This prevents smoothing of sharp-corners by discarding surface-normals. Useful for engineering CAD.
-            'sensor': {
-                'type': 'irradiancemeter',
-                'film': {
-                    'type': 'hdrfilm',
-                    'pixel_format': 'luminance',
-                    'filter': {'type': 'box'},
-                    'width': 1,
-                    'height': 1,
-                }
-            }
-        }
-
         t = time.time()
-        print('Loading Scene....')
-        scene = mi.load_dict(scene)
-        print('Time to load scene: ' + str(time.time() - t))
+        rt = shadowKernels()
+        scene = rt.buildMitsubaScene(self.intersectPoints, mitsubaMode=mitsubaMode)
+        print('Time to build Mitsuba scene: ' + str(time.time() - t))
         print('\n')
 
         t = time.time()
-        # Convert the numpy arrays to Dr.Jit dynamic arrays
         tx = dr.cuda.Float(self.sources[:,0]) if mitsubaMode == 'cuda' else dr.llvm.Float(self.sources[:,0])
         ty = dr.cuda.Float(self.sources[:,1]) if mitsubaMode == 'cuda' else dr.llvm.Float(self.sources[:,1])
         tz = dr.cuda.Float(self.sources[:,2]) if mitsubaMode == 'cuda' else dr.llvm.Float(self.sources[:,2])
         tp = dr.cuda.Float(self.sourcePower) if mitsubaMode == 'cuda' else dr.llvm.Float(self.sourcePower)
 
-
-        # Grab all the shape ids to correctly identify the sensor
-        idL = [x.id() for x in scene.shapes()]
-
-        print('Time to build sources: ' + str(time.time() - t))
-
-        t = time.time()
-        # Find the sensor
-        sensInd = idL.index(name)
-        sensS = scene.shapes()[sensInd]
-
-        params = mi.traverse(sensS)
-        verts = params['vertex_positions']
-        norms = params['vertex_normals']
-
-        # Find all the face and areas
-        totFace = sensS.face_count()
-        if mitsubaMode == 'cuda':
-            indF = sensS.face_indices(dr.arange(dr.cuda.ad.UInt, totFace))
-        else:
-            indF = sensS.face_indices(dr.arange(dr.llvm.ad.UInt, totFace))
-
-        # Label each vertex with its index
-        vertNum = int(len(verts) / 3)
-        vertA = np.array(verts)
-
-        vertAR = vertA.reshape((vertNum, 3))
-        indFA = np.array(indF)
-
-        # Find the vertices of each face
-        faceV = np.take(vertAR, indFA.flatten(), axis=0)
-
-        # Find the area of each face and the mean point of each face
-        faceV1 = faceV[::3, :]
-        faceV2 = faceV[1::3, :]
-        faceV3 = faceV[2::3, :]
-
-        # Find the area of each face
-        cross1 = faceV2 - faceV1
-        cross2 = faceV3 - faceV1
-
-        cross = np.cross(cross1, cross2)
-        area = np.linalg.norm(cross, axis=1) / 2
-
-        #find the mean point of each face
-        #again can probably be sped up
-        center = np.asarray([x.mean(0) for x in np.split(faceV,totFace,axis = 0)])
+        center, area = triangle_centroids_and_areas(self.intersectPoints)
+        totFace = center.shape[0]
 
         #dynamically allocate batch size based upon available memory
         if batch_size == "auto":
             batch_size = int(2**32 / totFace  * 0.5)
 
         print("Calculated batch size of: {:d}".format(batch_size))
-        
+
         N_sources = len(tx)
-        N_targets = len(center)
-        targetPower = np.zeros((N_targets))
+        targetPower = np.zeros((totFace))
         print('Time to prepare mesh: ' + str(time.time() - t))
 
         t0 = time.time()
@@ -824,6 +718,9 @@ class RAD:
         Maps power between sources and targets (ROI PFCs).  Uses Open3D to
         perform ray tracing.  Open3D can be optimized for CPU or GPU.
 
+        Mesh and ray origins use meters via rayTracerClass.shadowKernels.buildOpen3Dscene
+        on self.intersectPoints
+
         Uses Open3D to accelerate the calculation:
         Zhou, Qian-Yi, Jaesik Park, and Vladlen Koltun. "Open3D: A modern
         library for 3D data processing." arXiv preprint arXiv:1801.09847 (2018).
@@ -834,13 +731,10 @@ class RAD:
         Psum = np.zeros((self.Nj))
         #self.hullPower = np.zeros((self.Ni))
 
-        print("Building radiation scene...")
-        #build mesh and tensors for open3d
-        mesh = o3d.io.read_triangle_mesh(self.meshFile)
-        mesh.compute_vertex_normals()
-        mesh = o3d.t.geometry.TriangleMesh.from_legacy(mesh)
-        scene = o3d.t.geometry.RaycastingScene()
-        mesh_id = scene.add_triangles(mesh)
+        print("Building radiation scene (rayTracerClass, meters)...")
+        rt = shadowKernels()
+        rt.buildOpen3Dscene(np.asarray(self.intersectPoints, dtype=np.float32))
+        scene = rt.scene
         print("Scene building took {:f} [s]\n".format(time.time() - t0))
 
         #check how much memory we have
@@ -876,12 +770,12 @@ class RAD:
             r_ij = self.targetCtrs - self.sources[i]
             #r_ij *= 1000.0
             rMag = np.linalg.norm(r_ij, axis=1)
-            rNorm = r_ij / rMag.reshape((-1,1))
-            rdotn = np.sum(rNorm*self.targetNorms, axis=1)
-            q1 = np.tile(self.sources[i]*1000.0,(self.Nj,1))
+            rNorm = (r_ij / rMag.reshape((-1, 1))).astype(np.float32)
+            rdotn = np.sum(rNorm.astype(np.float64) * self.targetNorms, axis=1)
+            q1 = np.tile(self.sources[i], (self.Nj, 1)).astype(np.float32)
 
             #calculate intersections
-            rays = o3d.core.Tensor([np.hstack([q1,rNorm])],dtype=o3d.core.Dtype.Float32)
+            rays = o3d.core.Tensor([np.hstack([q1, rNorm])], dtype=o3d.core.Dtype.Float32)
             hits = scene.cast_rays(rays)
 
             #convert open3d CPU tensors back to numpy
